Added InputManager script based on online reference. Script enables 2D drag and drop functionality for both mouse and touch input.
Basic mouse input is tested and working, touch input is untested. A simple test object is now in the test scene, and it can be clicked and dragged around.
The input manager will need to be refined greatly as we develop other scripts before it can provide all its intended functionality.