Input Manager improved and redefined as a MoveController and Gesture scripts. The Draggable script manages object responses to input, such as drag or tap. The Block script manages the fundamental block behaviors while being interacted with, either dragged or tapped. Raycasting is being used with 3D physics to control when gravity is enabled on the blocks (only on when released in PlayArea). Blocks will return to initial location if released outside of the PlayArea. This function will later need to be refined to return Blocks to their positions within the Player Queue they were dragged from. Blocks also become undraggable once placed in the PlayArea.
A game manager script manages preload functions, player turns, and similar functions. The player turn does not change on an event yet. There is a button to toggle Player Turns in the Game Manager inspector. There are also some prefabs set up, most importantly the Block.

All changes are currently within the SampleScene. Player 1 Blocks are on the left, Player 2 Blocks are on the right. "Next Player" button must be used to switch players for debugging until collision events are implemented.